+++ packages/core/client/src/schema-component/antd/cascader/__tests__/cascader.test.tsx
@@ -1,5 +1,5 @@
-import { fireEvent, render, screen, sleep, userEvent } from '@nocobase/test/client';
 import React from 'react';
+import { fireEvent, render, screen, sleep, userEvent } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
