+++ packages/core/client/src/schema-component/antd/checkbox/__tests__/checkbox.test.tsx
@@ -1,5 +1,5 @@
-import { render, screen, userEvent } from '@nocobase/test/client';
 import React from 'react';
+import { render, screen, userEvent } from 'testUtils';
 import App1 from '../demos/checkbox';
 import App2 from '../demos/checkbox.group';
