+++ packages/core/client/src/schema-component/antd/date-picker/__tests__/date-picker.test.tsx
@@ -1,5 +1,5 @@
-import { render, screen, userEvent } from '@nocobase/test/client';
 import React from 'react';
+import { render, screen, userEvent } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
 import App3 from '../demos/demo3';
