+++ packages/core/client/src/schema-component/antd/filter/__tests__/filter.test.tsx
@@ -1,5 +1,5 @@
-import { render, screen, userEvent, within } from '@nocobase/test/client';
 import React from 'react';
+import { render, screen, userEvent, within } from 'testUtils';
 import App2 from '../demos/demo2';
 import App3 from '../demos/demo3';
 import App4 from '../demos/demo4';
