+++ packages/core/client/src/schema-component/antd/radio/__tests__/radio.test.tsx
@@ -1,5 +1,5 @@
-import { fireEvent, render, screen } from '@nocobase/test/client';
 import React from 'react';
+import { fireEvent, render, screen } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
 import App3 from '../demos/demo3';
