+++ packages/core/client/src/schema-component/antd/select/__tests__/select.test.tsx
@@ -1,5 +1,5 @@
-import { fireEvent, render, screen, userEvent } from '@nocobase/test/client';
 import React from 'react';
+import { fireEvent, render, screen, userEvent } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
 import App3 from '../demos/demo3';
