+++ packages/core/client/src/schema-component/antd/tree-select/__tests__/tree-select.test.tsx
@@ -1,5 +1,5 @@
-import { render, screen, userEvent } from '@nocobase/test/client';
 import React from 'react';
+import { render, screen, userEvent } from 'testUtils';
 import App1 from '../demos/demo1';
 
 describe('TreeSelect', () => {
