+++ packages/core/test/package.json
@@ -6,7 +6,6 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@nocobase/server": "0.9.4-alpha.2",
-    "@nocobase/client": "0.9.4-alpha.2",
     "@types/supertest": "^2.0.11",
     "mockjs": "^1.1.0",
     "mysql2": "^2.3.3",
