--- packages/core/test/src/client/index.ts
@@ -1,3 +1,0 @@
-export * from './mockAPIClient';
-export * from './utils';
-
